Move jest utils to root level test/utils
