+++ framework/test/jest/integration/specs/modules/chain/components/storage/entities/account/increase_field_by.spec.js
@@ -16,7 +16,7 @@ const {
 	Account,
 } = require('../../../../../../../../../../src/modules/chain/components/storage/entities');
 const { PgHelper } = require('../../../../../../../utils/pg-helper');
-const { constants } = require('../../../../../../../../utils');
+const { constants } = require('../../../../../../../../../utils');
 
 describe('storage.entities.Account.increaseFieldBy', () => {
 	let pgHelper;
