+++ framework/test/jest/integration/specs/modules/chain/components/storage/entities/account/update.spec.js
@@ -16,7 +16,7 @@ const {
 	Account,
 } = require('../../../../../../../../../../src/modules/chain/components/storage/entities');
 const { PgHelper } = require('../../../../../../../utils/pg-helper');
-const { constants } = require('../../../../../../../../utils');
+const { constants } = require('../../../../../../../../../utils');
 
 describe('storage.entities.Account.update', () => {
 	let pgHelper;
