+++ framework/test/jest/unit/specs/modules/chain/bft/bft.spec.js
@@ -34,7 +34,7 @@ const {
 	FORK_STATUS_DOUBLE_FORGING,
 } = require('../../../../../../../src/modules/chain/bft');
 
-const { constants } = require('../../../../../utils');
+const { constants } = require('../../../../../../utils');
 
 const generateBlocks = ({ startHeight, numberOfBlocks }) =>
 	new Array(numberOfBlocks)
