+++ framework/test/jest/unit/specs/modules/chain/synchronizer/block_synchronization_mechanism/block_synchronization_mechanism.spec.js
@@ -42,7 +42,7 @@ const {
 const {
 	Processor,
 } = require('../../../../../../../../src/modules/chain/processor');
-const { constants } = require('../../../../../../utils');
+const { constants } = require('../../../../../../../utils');
 const { newBlock } = require('../../../chain/blocks/utils');
 
 const genesisBlockDevnet = require('../../../../../../../fixtures/config/devnet/genesis_block');
