+++ framework/test/jest/unit/specs/modules/chain/synchronizer/fast_chain_switching_mechanism/fast_chain_switching_mechanism.spec.js
@@ -34,7 +34,7 @@ const {
 const {
 	Processor,
 } = require('../../../../../../../../src/modules/chain/processor');
-const { constants } = require('../../../../../../utils');
+const { constants } = require('../../../../../../../utils');
 const { newBlock } = require('../../../chain/blocks/utils');
 
 const genesisBlockDevnet = require('../../../../../../../fixtures/config/devnet/genesis_block');
