+++ framework/test/jest/unit/specs/modules/chain/synchronizer/synchronizer.spec.js
@@ -30,7 +30,7 @@ const {
 const {
 	Processor,
 } = require('../../../../../../../src/modules/chain/processor');
-const { constants } = require('../../../../../utils');
+const { constants } = require('../../../../../../utils');
 const { newBlock } = require('../../chain/blocks/utils');
 const synchronizerUtils = require('../../../../../../../src/modules/chain/synchronizer/utils');
