+++ framework/test/utils/deep_freeze.js
