+++ framework/test/utils/index.js
@@ -16,9 +16,7 @@
 
 const { constants } = require('./constants');
 
-const randomInt = (low, high) => {
-	return Math.round(Math.random() * (high - low) + low);
-};
+const randomInt = (low, high) => Math.round(Math.random() * (high - low) + low);
 
 module.exports = {
 	constants,
