+++ framework/test/utils/network_identifier.js
