+++ framework/test/utils/registered_transactions.js
